# add Chai's pathval features to this separate module