--- a/MAINTAINERS
+++ b/MAINTAINERS
@@ -1,0 +1,5 @@
+keithamus
+davelosert
+lucasfcosta
+meeber
+vesln