--- a/package.json
+++ b/package.json
@@ -1,27 +1,81 @@
 {
   "name": "pathval",
-  "version": "0.1.1",
   "description": "Object value retrieval given a string path",
-  "main": "./index.js",
   "homepage": "https://github.com/chaijs/pathval",
-  "scripts": {
-    "test": "hydro",
-    "coverage": "istanbul cover _hydro",
-    "test-readme": "jsmd README.md"
-  },
+  "keywords": [
+    "pathval",
+    "value retrieval",
+    "chai util"
+  ],
+  "license": "MIT",
+  "author": "Veselin Todorov <hi@vesln.com>",
+  "files": [
+    "index.js",
+    "pathval.js"
+  ],
+  "main": "./index.js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/chaijs/pathval"
+    "url": "git+ssh://git@github.com/chaijs/pathval.git"
+  },
+  "scripts": {
+    "build": "browserify --bare $npm_package_main --standalone pathval -o pathval.js",
+    "lint": "eslint --ignore-path .gitignore .",
+    "prepublish": "npm run build",
+    "semantic-release": "semantic-release pre && npm publish && semantic-release post",
+    "pretest": "npm run lint",
+    "test": "npm run test:node && npm run test:browser && npm run upload-coverage",
+    "test:browser": "karma start --singleRun=true",
+    "test:node": "istanbul cover _mocha",
+    "upload-coverage": "lcov-result-merger 'coverage/**/lcov.info' | coveralls; exit 0"
+  },
+  "config": {
+    "ghooks": {
+      "commit-msg": "validate-commit-msg"
+    }
+  },
+  "eslintConfig": {
+    "extends": [
+      "strict/es5"
+    ],
+    "env": {
+      "es6": true
+    },
+    "globals": {
+      "HTMLElement": false
+    },
+    "rules": {
+      "complexity": 0,
+      "max-statements": 0
+    }
+  },
+  "dependencies": {
+    "type-detect": "^2.0.1"
   },
   "devDependencies": {
-    "istanbul": "~0.4.3",
-    "jsmd": "~0.2.0",
-    "simple-assert": "~1.0.0",
-    "hydro": "*",
-    "hydro-file-suite": "*",
-    "hydro-doc": "*",
-    "hydro-bdd": "*"
+    "browserify": "^13.0.0",
+    "browserify-istanbul": "^1.0.0",
+    "coveralls": "2.11.9",
+    "eslint": "^2.4.0",
+    "eslint-config-strict": "^8.5.0",
+    "eslint-plugin-filenames": "^0.2.0",
+    "ghooks": "^1.0.1",
+    "istanbul": "^0.4.2",
+    "karma": "^0.13.22",
+    "karma-browserify": "^5.0.2",
+    "karma-coverage": "^0.5.5",
+    "karma-mocha": "^0.2.2",
+    "karma-phantomjs-launcher": "^1.0.0",
+    "karma-sauce-launcher": "^0.3.1",
+    "lcov-result-merger": "^1.0.2",
+    "mocha": "^2.4.5",
+    "phantomjs-prebuilt": "^2.1.5",
+    "semantic-release": "^4.3.5",
+    "simple-assert": "^1.0.0",
+    "travis-after-all": "^1.4.4",
+    "validate-commit-msg": "^2.3.1"
   },
-  "author": "Veselin Todorov <hi@vesln.com>",
-  "license": "MIT"
+  "engines": {
+    "node": "*"
+  }
 }