This PR holds the progress I've made on moving Chai's `pathval` features to this separate module.
I have done some changes to the code brought here in order to have 100% coverage and follow the linting rules defined.

This is almost done, but I'd still like to refactor some parts of the code, such as turning the `internalSetPathValue` function into a recursive function and maybe renaming some variables for them to make sense.

If you have any other refactoring suggestions please let me know.
I'll remove the `[WIP]` tag and do another post here when I finish this one.

**Important Detail:**
[Currently the build is failing for Node `0.8` because `type-detect` (which is a dependency of `pathval`) needs a Node version above `0.1.0`.](https://travis-ci.org/chaijs/pathval/jobs/141903679#L230) What are we going to do about it? According to [the `type-detect` package.json](https://github.com/chaijs/type-detect/blob/master/package.json#L83) it should be compatible with any Node.js version as soon as the version on the repo gets released.

Thanks for your attention 😄 

---------------------------------------------------------------------
EDIT: This one is done. Please read the comment below.