@lucasfcosta 2016-07-09
Hey everyone, this one is done!
I thought it would be good to refactor `getPathValue` and make it a recursive method instead of using a `for` loop, but after quite some time thinking about it I thought it would be better not to mess with it because we've gotta track the state of the properties we need and this means we gotta have an index for then, so passing the same `parsed` object around wouldn't do much good and would only add unnecessary complexity to the code. If you guys really think this is necessary let me know, but I think what we've got for now is sufficient.

I would also like to ask you if I should add two methods called `.get` and `.set` that call `getPathValue` and `setPathValue` in order not to break backwards compatibility with the old API.

Please let me know if you find any problems with this

PS.: Please read the note above about the NodeJS version required for this package