@lucasfcosta 2016-07-09
Now it's really done!
This is a really big one, so let me know if I missed any typos or anything.
Thanks for the help @meeber 😄