@meeber 2016-07-09
@lucasfcosta Looks like you've been busy lately. Great work! I think this module has a lot of value being standalone; the string syntax of property access is pretty nifty :D 

In addition to the line comments above, I have a couple questions:

- I find the parameter ordering a bit surprising in many of the functions. In particular, I'd expect `obj` to be the first parameter instead of the last. It looks like `obj` did use to be first in `.set`. and `.get`. Was the change intentional?
- What's with `test.sh`?