@lucasfcosta 2016-07-09
@meeber Yeah, I took those functions out of Chai core and they had this parameter order, so I thought it should've been kept the same, that's why [I had the doubt if I should or add `get` and `set` methods with the same signatures as before just to keep the API compatible with the previous version](https://github.com/chaijs/pathval/pull/11#issuecomment-231501186).

`test.sh` is a test file I was using in order to run tests on my IE VMs and make sure everything was working on IE too, I better remove that, well noticed.

I have also added a comment explaining why `.defined(val)` exists because I also thought it keeps the code cleaner, it's better to write `defined(val)` at a lot of places than having to use `typeof` every single time. What do you think about this?

Thanks for reviewing, it's always good to have another pair of eyes on such a big PR.