@meeber 2016-07-09
@lucasfcosta Adding the `get` and `set` methods like before would be good. If this were a new project, I'd say all of the methods should accept `obj` as their first parameter, but it's probably not worth introducing such a breaking change after the fact.