@keithamus 2016-07-12
Having all of the methods take the object first seems sensible, right now they're inconsistent and thats, IMO, the worst case scenario. 

Remember this (and chai) will be a breaking version so - while we should make the effort to ensure this doesn't break lots of plugins - we also have license to change things if it makes the libs better, which I think unequivocally it does here.