@lucasfcosta 2016-07-14
Here it is. Thanks for your input guys 😄 
I changed all the methods to take the `object` first and then the `path`.
I also updated all the docs (including `README.md`) and the docStrings' examples

I also removed some unnecessary `defined` checks [as noticed by @meeber here](https://github.com/chaijs/pathval/pull/11#discussion_r70160880).

Since @keithamus said this will be a breaking version I didn't add the old `set` and `get` methods, the ones on this PR have the same behavior as the old ones even though they have different names.

I think this is good to go, please let me know if you find anything else that needs to be changed.