@meeber 2016-07-16
@lucasfcosta Having all of the methods take `object` as their first parameter brings me far greater happiness than such a small thing ever should.

This looks just about done to me. I'm fine with no `set` and `get` methods given the available equivalents.

There does seem to be one outstanding item regarding @keithamus' note about benchmarking the `defined` helper function.