@lucasfcosta 2016-07-17
Hello everyone, well I've used [this module](https://www.npmjs.com/package/benchmark) and I have created two versions of the `internalSetPathValue` and `setPathValue` to measure the performance of the `defined` method and `typeof obj !== 'undefined' and I have discovered that `defined` is a little bit faster than using `typeof`. This is not a much significant difference and one time the `typeof version was a little bit faster.

Here are the results I have collected:

```
# Result 1
Without defined() x 278,315 ops/sec ±11.08% (67 runs sampled)
With defined() x 280,066 ops/sec ±10.67% (72 runs sampled)

# Result 2
Without defined() x 280,427 ops/sec ±9.83% (70 runs sampled)
With defined() x 293,800 ops/sec ±7.80% (77 runs sampled)

# Result 3
Without defined() x 280,420 ops/sec ±8.44% (74 runs sampled)
With defined() x 278,014 ops/sec ±8.95% (75 runs sampled)

# Result 4
Without defined() x 231,060 ops/sec ±14.99% (60 runs sampled)
With defined() x 267,356 ops/sec ±9.32% (67 runs sampled)
```

I also ran this a couple more times and got even less significant results, these are the four first executions, which seem to summarize the situation pretty well.

So, as @keithamus suggested I'm going to drop this method.

Thank you guys for making me dig deeper in this. This reminds me of [the first of the Akin's Laws of
Spacecraft Design](http://spacecraft.ssl.umd.edu/akins_laws.html), which states that:

> 1. Engineering is done with numbers. Analysis without numbers is only an opinion.

Well, that was quite a lesson to learn today, have a nice sunday folks 😄