@lucasfcosta 2016-07-25
Ping @keithamus @meeber 
Hey guys, I'm sorry to bother but I'll be going to spend my vacation abroad so I may not be able to commit between July 28th and Aug 12th, so if you guys need anything regarding this one (or any other thing) please let me know on this PR (or tag me in the issue you need me) and I'll do my best to get things going again as soon as I get back from the trip. 😄 

However, until July 28th, you can count on me 👍