@meeber 2016-07-25
@lucasfcosta Vacation? I don't remember approving a vacation.

@keithamus Did you approve this vacation?

Regarding this PR: LGTM