@lucasfcosta 2016-07-26
@meeber Oh I'm sorry Mr. Meeber, I talked with the Internet's Human Resources department and they said it was okay.
I hope I could send you some pizza straight from Italy but I doubt it will get to Virginia before it gets too cold.