@keithamus 2016-08-01
Okay, sorry for holding this one up. LGTM. Let's merge it and get working on releasing 😄