@keithamus 2016-08-01
Also, I most certainly did not approve of any vacation! Our vacation policy is very strict: you're only allowed a max of 365 days a year and all vacation will go unpaid.