@CacheControl 2016-09-11
@lucasfcosta @keithamus 

Can you provide an update on any plans for publishing a new major version?  As this repo currently stands, it's very confusing to `npm install` and receive as a completely different api than what the github README shows (e.g. `pathvalue.set()` exists instead of `pathvalue. setPathValue()`

On that note, was there any particular reason to rename `.get()` and `.set()`?  If the functionality hasn't changed, it'd be nice to at least alias those methods to their new counterparts for providing backwards compatibility with the previous api.