@keithamus 2016-09-11
Hey @CacheControl. You can check the website for the docs for the latest version of chai. We'll be releasing 4.0 pretty soon, and it'll be a major breaking version, which means api changes like what you've described will happen. We will be releasing comprehensive changelogs with the release of 4.0 so you know what has changed.